Crate zcash_encoding

Zcash binary encodings.

zcash_encoding is a library that provides common encoding and decoding operations for stable binary encodings used throughout the Zcash ecosystem.

Structs

Array

Namespace for functions that perform encoding of array contents.

CompactSize

Namespace for functions for compact encoding of integers.

Optional

Namespace for functions that perform encoding of Option values.

Vector

Namespace for functions that perform encoding of vectors.

Constants

MAX_COMPACT_SIZE

The maximum allowed value representable as a [CompactSize]